Adds context hash and translation_key to errors
The context hash contains additional information pertaining to the error.
Of primary concern is the property value which generally refers to the specific json object property that is failing validation.
All errors now also return a translation key prefixed with 'json_schema_error_'.  These are intended to be used with a translation library like I18n.